+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - events deleting
+- static members deleting
 
 ## [0.6.1] - 2019-11-04
 ### Added